Used custom hook to fetch content
